+++ src/base/widget.c
@@ -2667,6 +2667,16 @@ uint32_t widget_add_timer(widget_t* widget, timer_func_t on_timer, uint32_t dura
   return id;
 }
 
+ret_t widget_remove_timer(widget_t* widget, uint32_t timer_id) {
+  return_value_if_fail(widget != NULL, TK_INVALID_ID);
+
+  if (widget_off_by_func(widget, EVT_DESTROY, widget_remove_timer_on_destroy, tk_pointer_from_int(timer_id)) == RET_OK) {
+    return timer_remove(timer_id);
+  }
+
+  return RET_BAD_PARAMS;
+}
+
 static ret_t widget_remove_idle_on_destroy(void* ctx, event_t* e) {
   uint32_t id = (char*)ctx - (char*)NULL;
   idle_remove(id);